Move all create subcommands to its own subdirectory
