+++ hack/.golint_failures
@@ -139,6 +139,7 @@ pkg/kubectl/categories
 pkg/kubectl/cmd
 pkg/kubectl/cmd/auth
 pkg/kubectl/cmd/config
+pkg/kubectl/cmd/create
 pkg/kubectl/cmd/rollout
 pkg/kubectl/cmd/set
 pkg/kubectl/cmd/templates
