+++ pkg/kubectl/cmd/BUILD
@@ -24,21 +24,6 @@ go_library(
         "completion.go",
         "convert.go",
         "cp.go",
-        "create.go",
-        "create_clusterrole.go",
-        "create_clusterrolebinding.go",
-        "create_configmap.go",
-        "create_deployment.go",
-        "create_job.go",
-        "create_namespace.go",
-        "create_pdb.go",
-        "create_priorityclass.go",
-        "create_quota.go",
-        "create_role.go",
-        "create_rolebinding.go",
-        "create_secret.go",
-        "create_service.go",
-        "create_serviceaccount.go",
         "delete.go",
         "describe.go",
         "diff.go",
@@ -82,6 +67,7 @@ go_library(
         "//pkg/kubectl/apply/strategy:go_default_library",
         "//pkg/kubectl/cmd/auth:go_default_library",
         "//pkg/kubectl/cmd/config:go_default_library",
+        "//pkg/kubectl/cmd/create:go_default_library",
         "//pkg/kubectl/cmd/resource:go_default_library",
         "//pkg/kubectl/cmd/rollout:go_default_library",
         "//pkg/kubectl/cmd/scalejob:go_default_library",
@@ -114,11 +100,8 @@ go_library(
         "//vendor/github.com/renstrom/dedent:go_default_library",
         "//vendor/github.com/spf13/cobra:go_default_library",
         "//vendor/github.com/spf13/pflag:go_default_library",
-        "//vendor/k8s.io/api/batch/v1:go_default_library",
-        "//vendor/k8s.io/api/batch/v1beta1:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
         "//vendor/k8s.io/api/policy/v1beta1:go_default_library",
-        "//vendor/k8s.io/api/rbac/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/errors:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/meta:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
@@ -144,9 +127,7 @@ go_library(
         "//vendor/k8s.io/apiserver/pkg/util/flag:go_default_library",
         "//vendor/k8s.io/client-go/discovery:go_default_library",
         "//vendor/k8s.io/client-go/kubernetes:go_default_library",
-        "//vendor/k8s.io/client-go/kubernetes/typed/batch/v1:go_default_library",
         "//vendor/k8s.io/client-go/kubernetes/typed/core/v1:go_default_library",
-        "//vendor/k8s.io/client-go/kubernetes/typed/rbac/v1:go_default_library",
         "//vendor/k8s.io/client-go/rest:go_default_library",
         "//vendor/k8s.io/client-go/scale:go_default_library",
         "//vendor/k8s.io/client-go/tools/clientcmd:go_default_library",
@@ -171,21 +152,6 @@ go_test(
         "cmd_test.go",
         "convert_test.go",
         "cp_test.go",
-        "create_clusterrole_test.go",
-        "create_clusterrolebinding_test.go",
-        "create_configmap_test.go",
-        "create_deployment_test.go",
-        "create_job_test.go",
-        "create_namespace_test.go",
-        "create_pdb_test.go",
-        "create_priorityclass_test.go",
-        "create_quota_test.go",
-        "create_role_test.go",
-        "create_rolebinding_test.go",
-        "create_secret_test.go",
-        "create_service_test.go",
-        "create_serviceaccount_test.go",
-        "create_test.go",
         "delete_test.go",
         "describe_test.go",
         "diff_test.go",
@@ -222,6 +188,7 @@ go_test(
         "//pkg/apis/core:go_default_library",
         "//pkg/apis/extensions:go_default_library",
         "//pkg/kubectl:go_default_library",
+        "//pkg/kubectl/cmd/create:go_default_library",
         "//pkg/kubectl/cmd/resource:go_default_library",
         "//pkg/kubectl/cmd/testing:go_default_library",
         "//pkg/kubectl/cmd/util:go_default_library",
@@ -235,16 +202,10 @@ go_test(
         "//pkg/util/strings:go_default_library",
         "//vendor/github.com/googleapis/gnostic/OpenAPIv2:go_default_library",
         "//vendor/github.com/spf13/cobra:go_default_library",
-        "//vendor/github.com/stretchr/testify/assert:go_default_library",
         "//vendor/gopkg.in/yaml.v2:go_default_library",
         "//vendor/k8s.io/api/autoscaling/v1:go_default_library",
-        "//vendor/k8s.io/api/batch/v1:go_default_library",
-        "//vendor/k8s.io/api/batch/v1beta1:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
         "//vendor/k8s.io/api/policy/v1beta1:go_default_library",
-        "//vendor/k8s.io/api/rbac/v1:go_default_library",
-        "//vendor/k8s.io/api/rbac/v1beta1:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/api/equality:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/errors:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/meta:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/resource:go_default_library",
@@ -261,7 +222,6 @@ go_test(
         "//vendor/k8s.io/apimachinery/pkg/util/wait:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/version:go_default_library",
         "//vendor/k8s.io/client-go/dynamic:go_default_library",
-        "//vendor/k8s.io/client-go/kubernetes/fake:go_default_library",
         "//vendor/k8s.io/client-go/rest:go_default_library",
         "//vendor/k8s.io/client-go/rest/fake:go_default_library",
         "//vendor/k8s.io/client-go/scale/fake:go_default_library",
@@ -286,6 +246,7 @@ filegroup(
         ":package-srcs",
         "//pkg/kubectl/cmd/auth:all-srcs",
         "//pkg/kubectl/cmd/config:all-srcs",
+        "//pkg/kubectl/cmd/create:all-srcs",
         "//pkg/kubectl/cmd/resource:all-srcs",
         "//pkg/kubectl/cmd/rollout:all-srcs",
         "//pkg/kubectl/cmd/scalejob:all-srcs",
