+++ pkg/kubectl/cmd/create/create_clusterrolebinding_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package cmd
+package create
 
 import (
 	"bytes"
@@ -24,7 +24,7 @@ import (
 	"reflect"
 	"testing"
 
-	rbac "k8s.io/api/rbac/v1"
+	rbac "k8s.io/api/rbac/v1beta1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
@@ -34,21 +34,19 @@ import (
 	cmdtesting "k8s.io/kubernetes/pkg/kubectl/cmd/testing"
 )
 
-var groupVersion = schema.GroupVersion{Group: "rbac.authorization.k8s.io", Version: "v1"}
-
-func TestCreateRoleBinding(t *testing.T) {
-	expectBinding := &rbac.RoleBinding{
-		TypeMeta: v1.TypeMeta{
-			APIVersion: "rbac.authorization.k8s.io/v1",
-			Kind:       "RoleBinding",
-		},
+func TestCreateClusterRoleBinding(t *testing.T) {
+	expectBinding := &rbac.ClusterRoleBinding{
 		ObjectMeta: v1.ObjectMeta{
 			Name: "fake-binding",
 		},
+		TypeMeta: v1.TypeMeta{
+			Kind:       "ClusterRoleBinding",
+			APIVersion: "rbac.authorization.k8s.io/v1beta1",
+		},
 		RoleRef: rbac.RoleRef{
 			APIGroup: rbac.GroupName,
-			Kind:     "Role",
-			Name:     "fake-role",
+			Kind:     "ClusterRole",
+			Name:     "fake-clusterrole",
 		},
 		Subjects: []rbac.Subject{
 			{
@@ -79,29 +77,29 @@ func TestCreateRoleBinding(t *testing.T) {
 	decoder := ns.DecoderToVersion(info.Serializer, groupVersion)
 
 	tf.Namespace = "test"
-	tf.Client = &RoleBindingRESTClient{
+	tf.Client = &ClusterRoleBindingRESTClient{
 		RESTClient: &fake.RESTClient{
 			NegotiatedSerializer: ns,
 			Client: fake.CreateHTTPClient(func(req *http.Request) (*http.Response, error) {
 				switch p, m := req.URL.Path, req.Method; {
-				case p == "/namespaces/test/rolebindings" && m == "POST":
+				case p == "/clusterrolebindings" && m == "POST":
 					bodyBits, err := ioutil.ReadAll(req.Body)
 					if err != nil {
-						t.Fatalf("TestCreateRoleBinding error: %v", err)
+						t.Fatalf("TestCreateClusterRoleBinding error: %v", err)
 						return nil, nil
 					}
 
-					if obj, _, err := decoder.Decode(bodyBits, nil, &rbac.RoleBinding{}); err == nil {
-						if !reflect.DeepEqual(obj.(*rbac.RoleBinding), expectBinding) {
-							t.Fatalf("TestCreateRoleBinding: expected:\n%#v\nsaw:\n%#v", expectBinding, obj.(*rbac.RoleBinding))
+					if obj, _, err := decoder.Decode(bodyBits, nil, &rbac.ClusterRoleBinding{}); err == nil {
+						if !reflect.DeepEqual(obj.(*rbac.ClusterRoleBinding), expectBinding) {
+							t.Fatalf("TestCreateClusterRoleBinding: expected:\n%#v\nsaw:\n%#v", expectBinding, obj.(*rbac.ClusterRoleBinding))
 							return nil, nil
 						}
 					} else {
-						t.Fatalf("TestCreateRoleBinding error, could not decode the request body into rbac.RoleBinding object: %v", err)
+						t.Fatalf("TestCreateClusterRoleBinding error, could not decode the request body into rbac.ClusterRoleBinding object: %v", err)
 						return nil, nil
 					}
 
-					responseBinding := &rbac.RoleBinding{}
+					responseBinding := &rbac.ClusterRoleBinding{}
 					responseBinding.Name = "fake-binding"
 					return &http.Response{StatusCode: 201, Header: defaultHeader(), Body: ioutil.NopCloser(bytes.NewReader([]byte(runtime.EncodeOrDie(encoder, responseBinding))))}, nil
 				default:
@@ -112,34 +110,44 @@ func TestCreateRoleBinding(t *testing.T) {
 		},
 	}
 
+	expectedOutput := "clusterrolebinding.rbac.authorization.k8s.io/" + expectBinding.Name + "\n"
 	buf := bytes.NewBuffer([]byte{})
-	cmd := NewCmdCreateRoleBinding(tf, buf)
-	cmd.Flags().Set("role", "fake-role")
+	cmd := NewCmdCreateClusterRoleBinding(tf, buf)
+	cmd.Flags().Set("clusterrole", "fake-clusterrole")
 	cmd.Flags().Set("user", "fake-user")
 	cmd.Flags().Set("group", "fake-group")
+	cmd.Flags().Set("output", "name")
 	cmd.Flags().Set("serviceaccount", "fake-namespace:fake-account")
 	cmd.Run(cmd, []string{"fake-binding"})
+	if buf.String() != expectedOutput {
+		t.Errorf("TestCreateClusterRoleBinding: expected %v\n but got %v\n", expectedOutput, buf.String())
+	}
 }
 
-type RoleBindingRESTClient struct {
+type ClusterRoleBindingRESTClient struct {
 	*fake.RESTClient
 }
 
-func (c *RoleBindingRESTClient) Post() *restclient.Request {
+func (c *ClusterRoleBindingRESTClient) Post() *restclient.Request {
 	config := restclient.ContentConfig{
 		ContentType:          runtime.ContentTypeJSON,
-		GroupVersion:         &groupVersion,
 		NegotiatedSerializer: c.NegotiatedSerializer,
 	}
 
 	info, _ := runtime.SerializerInfoForMediaType(c.NegotiatedSerializer.SupportedMediaTypes(), runtime.ContentTypeJSON)
 	serializers := restclient.Serializers{
-		Encoder: c.NegotiatedSerializer.EncoderForVersion(info.Serializer, groupVersion),
-		Decoder: c.NegotiatedSerializer.DecoderToVersion(info.Serializer, groupVersion),
+		Encoder: c.NegotiatedSerializer.EncoderForVersion(info.Serializer, schema.GroupVersion{Group: "rbac.authorization.k8s.io", Version: "v1beta1"}),
+		Decoder: c.NegotiatedSerializer.DecoderToVersion(info.Serializer, schema.GroupVersion{Group: "rbac.authorization.k8s.io", Version: "v1beta1"}),
 	}
 	if info.StreamSerializer != nil {
 		serializers.StreamingSerializer = info.StreamSerializer.Serializer
 		serializers.Framer = info.StreamSerializer.Framer
 	}
 	return restclient.NewRequest(c, "POST", &url.URL{Host: "localhost"}, c.VersionedAPIPath, config, serializers, nil, nil, 0)
+}
+
+func defaultHeader() http.Header {
+	header := http.Header{}
+	header.Set("Content-Type", runtime.ContentTypeJSON)
+	return header
 }
